requirements.txt
discord.py
openai
mysql-connector-python
requests
beautifulsoup4
html5lib
python-dotenv
